Fix Use-after-Free in qembeddingbag_byte_prepack_out
When FBGEMM is not used (either manually disabled or on platforms such
as POWER where it isn't supported at all) the fallback code requests a
`data_ptr<float>` on a `Tensor` object returned by
`to(ScalarType::Float)` in the same line. This object will be destroyed
at the end of the line leading to a dangling pointer.

On some platforms this manifests in wrong results being returned as the
memory gets overwritten.

Fix this by binding the temporary object (or initial object) to a const
value reference which extents its lifetime and getting the `data_ptr`
from that.

Fixes #84748