Fix MultiProcess failure on nodes with 1 GPU
The decorator(s) is written to `sys.exit` when the test function is
called which is AFTER the `setup` call which forks the processes and
uses (potentially) a GPU/NCCL based barrier which requires "n GPUs" to
be present befor checking if "n GPUs" are available.

Rewrite those decorators to use `unittest.skipIf` which will not even
enter the `setup` function.
This also exposed that `require_n_gpus_for_nccl_backend` is the same as
`nccl_skip_if_lt_x_gpu` but the former has a better name so I removed
the latter.

Fixes #89686