Allow mp.start_processes to create processes in parallel (#133707)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/133707

This is to fix the pytorch issue filed https://github.com/pytorch/pytorch/issues/133010

one way to fix this problem is to enable parallel start processes in mp.start_processes.

What else in the diff:
refactored a test case api_test which was repeating a lot of tests due to the inheritance.
added unit test for forkserver when parallel start is on.

Test Plan:
Added unit test.
https://fburl.com/testinfra/evuie3eh, all pass and the parallel forkserver version is sometimes faster.

Differential Revision: D61301603