[torchelastic] Don't do signal handling when off the main thread (#135088)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/135088

In multiprocessing, signal handling is not possible if the thread is not the main thread. This resulted in the following error:
> "ValueError('signal only works in main thread of the main interpreter')"

To address this issue, the diff checks whether the thread is the main thread and, if not, skips signal handling.

Test Plan:
Added a new unit test ```test_pcontext_wait_on_a_child_thread```
  buck2 test mode/dev-nosan //caffe2/test/distributed/elastic/multiprocessing:api_test -- --exact 'caffe2/test/distributed/elastic/multiprocessing:api_test - test_pcontext_wait_on_a_child_thread (api_test.ForkServerTest)'

Without the change:
Test failed with
> ValueError: signal only works in main thread of the main interpreter

With the change:
Test passed

Differential Revision: D62166943