Different file design, [export][refactor] create a folder that holds different major version schemas
Summary:
A lot of places are directly referencing the the dataclasses in schema.py. Given we need to keep the original data class to maintain BC after a major version bump, we'll create different schemas of different major versions.

Note that this also implies that we cannot remove a dataclass (or rename) in a major version bump.

Test Plan: Existing export tests.

Differential Revision: D71936455