Optimize the pytree utility module
This commit attempts to optimize the speed of the pytree module,
specifically the tree_is_leaf, tree_flatten, and tree_unflatten
functions.

The main optimization is that TreeSpec no longer functions as a
tree structure, but instead stores a flat list of nodes that are
generated by a post-order traversal of the tree. This speeds up
flattening a bit because we can use pointers into the leaf and
nodes list to calculate the subtree sizes, and drastically speeds
up unflattening because it becomes a linear loop through a list.

Most of the other changes in this commit follow from the structure
change of the TreeSpec class. This especially includes the
serialization functionality, which has a new schema. The various
serialization functions have been updated to handle dumping and
loading from either schema (as presumably the old schema will need
to stick around for a while for migration purposes).

This commit purposefully tries to avoid changing anything outside
of the _pytree.py file itself, which means there are plenty of
places where additional code could be deleted/changed to make it
make more sense. For example, the LeafSpec class no longer makes a
ton of sense, and is a good candidate for deletion in the future.

The overall results include ~3x improvement for tree_flatten, ~2x
improvement for tree_unflatten, and ~2x improvement for tree_map.
Full results are below:

```
tree_flatten:
    _cxx_pytree          5730.63 ips
    _pytree2             2439.90 ips  (-57.42%)
    _pytree              850.49 ips  (-85.16%)
tree_unflatten:
    _cxx_pytree          11241.22 ips
    _pytree2             4035.69 ips  (-64.10%)
    _pytree              1862.81 ips  (-83.43%)
tree_map:
    _cxx_pytree          3049.11 ips
    _pytree2             1100.40 ips  (-63.91%)
    _pytree              520.17 ips  (-82.94%)
tree_leaves:
    _cxx_pytree          5736.86 ips
    _pytree2             1736.03 ips  (-69.74%)
    _pytree              1525.42 ips  (-73.41%)
```