control_plane: add handler for WaitCounters (#167871)
Summary:

This adds a DebugServer handler for WaitCounters such that we can access all wait counters live via HTTP.

To do so we register a wait counter backend that tracks all counter values in a shared synchronized map. When creating a counter this will acquire the lock to add it to the global map but during runtime it only uses atomic operation.

Test Plan:
```
//caffe2/test/distributed/elastic:test_control_plane
```

Differential Revision: D87095718