Add check_lowerbound config for AOTI lowering
Summary:
## Why:
PT has a 0/1 specialization assumption. Thus, unless backed_size_oblivious is on, the greatest lowerbound is [2+, ...]. So, this trips up a lot of models who can have a dynamic size with 0 or 1.

I want a way for AOTI lowering users to easily check their dynamic shape spec is correct via `AOTI_RUNTIME_CHECK_INPUTS=1` which is a runtime env var. In-order to do that, I need to avoid any errors triggered by a wrong lowerbound.

## What does it look like?
```
    // this is a lowerbound check
    if (arg0_1_size[0] < 2) {
        std::stringstream ss;
        ss << "input_handles[0]: dim value is too small at 0, "
           << "expected it to be >= 2, " << "but got: "
           << arg0_1_size[0] << "\n";
        throw std::runtime_error(ss.str());
    }
```

Differential Revision: D88203028