Optimize tensor broadcasting (#170772)
Summary:

Continuing the trend of "Sympy is very expensive", there are a couple very common cases where we're doing unnecessary work to determine broadcasting and we should instead just short-circuit.

1) There is only one Tensor. In that case there is nothing to broadcast and we should just return its shape.

2) In `broadcast_symbolic_shapes` if `a` is a scalar or identical to `b` the answer is trivially b's shape.

3) If a Tensor is already the broadcasted shape, skip the (somewhat expensive) process of expanding it.

These degenerate cases account for the vast majority of broadcasting time in the max-autotune case that I'm looking at, and this change shaves ~5% off of `GraphLowering.run`.

Test Plan: Same as prior

Reviewed By: eellison, PaulZhang12

Differential Revision: D89199580